# resources/api/oresamsub-v2.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "OresamSub Business API",
    "version": "2.0.0",
    "description": "A single integration for data, airtime, cable TV and electricity reseller websites. All amounts are expressed in Nigerian naira."
  },
  "servers": [{"url": "https://oresamsub.com/api/v2", "description": "Production"}],
  "security": [{"bearerAuth": []}],
  "paths": {
    "/catalogue": {
      "get": {
        "summary": "List available plans",
        "operationId": "listCatalogue",
        "responses": {"200": {"description": "Plans priced for the authenticated business"}, "401": {"$ref": "#/components/responses/Unauthenticated"}}
      }
    },
    "/wallet": {
      "get": {
        "summary": "Get wallet balance",
        "operationId": "getWallet",
        "responses": {"200": {"description": "Available main-wallet balance"}, "401": {"$ref": "#/components/responses/Unauthenticated"}}
      }
    },
    "/validate-customer": {
      "post": {
        "summary": "Validate a cable or electricity customer",
        "operationId": "validateCustomer",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/ValidationRequest"}, "examples": {
            "cable": {"value": {"service": "cable", "plan_id": 2101, "customer": "1234567890"}},
            "electricity": {"value": {"service": "electricity", "plan_id": 3101, "customer": "01234567890"}}
          }}}
        },
        "responses": {"200": {"description": "Customer details and a validation reference valid for 10 minutes"}, "422": {"description": "Customer could not be validated"}}
      }
    },
    "/buy-service": {
      "post": {
        "summary": "Purchase data or airtime",
        "operationId": "buyService",
        "requestBody": {
          "required": true,
          "content": {"application/json": {"schema": {"$ref": "#/components/schemas/PurchaseRequest"}, "examples": {
            "data": {"value": {"service": "data", "plan_id": 1201, "customer": "08030000000", "reference": "ORDER-10001"}},
            "airtime": {"value": {"service": "airtime", "plan_id": 1401, "customer": "08030000000", "amount": 1000, "reference": "ORDER-10002"}},
            "cable": {"value": {"service": "cable", "plan_id": 2101, "customer": "1234567890", "validation_reference": "VAL-EXAMPLE", "reference": "ORDER-10003"}},
            "electricity": {"value": {"service": "electricity", "plan_id": 3101, "customer": "01234567890", "amount": 5000, "validation_reference": "VAL-EXAMPLE", "reference": "ORDER-10004"}}
          }}}
        },
        "responses": {
          "200": {"description": "Purchase processed or safely replayed"},
          "202": {"description": "Purchase is pending or processing"},
          "409": {"description": "Reference was previously used for different purchase details"},
          "422": {"description": "Invalid request or failed purchase"}
        }
      }
    },
    "/transactions/{reference}": {
      "get": {
        "summary": "Reconcile a transaction",
        "operationId": "getTransaction",
        "parameters": [{"name": "reference", "in": "path", "required": true, "schema": {"type": "string"}}],
        "responses": {"200": {"description": "Transaction owned by the authenticated business"}, "404": {"description": "Transaction not found"}}
      }
    }
  },
  "components": {
    "securitySchemes": {"bearerAuth": {"type": "http", "scheme": "bearer", "description": "API token generated from the OresamSub account"}},
    "responses": {"Unauthenticated": {"description": "Missing or invalid API token"}},
    "schemas": {
      "PurchaseRequest": {
        "type": "object",
        "required": ["service", "plan_id", "customer", "reference"],
        "properties": {
          "service": {"type": "string", "enum": ["data", "airtime", "cable", "electricity"]},
          "plan_id": {"description": "Plan ID returned by the catalogue endpoint"},
          "customer": {"type": "string", "pattern": "^0[789][01][0-9]{8}$"},
          "reference": {"type": "string", "maxLength": 100},
          "amount": {"type": "number", "minimum": 50, "description": "Required for airtime and electricity"},
          "validation_reference": {"type": "string", "description": "Required for cable and electricity; obtained from /validate-customer"},
          "validate_phone_network": {"type": "boolean", "default": true}
        }
      },
      "ValidationRequest": {
        "type": "object",
        "required": ["service", "plan_id", "customer"],
        "properties": {
          "service": {"type": "string", "enum": ["cable", "electricity"]},
          "plan_id": {"description": "Plan ID returned by the catalogue endpoint"},
          "customer": {"type": "string", "description": "Smartcard/IUC number for cable or meter number for electricity"}
        }
      }
    }
  }
}